frontend-data-broker/requirements.txt
flask
mysql-connector-python
python-dotenv
requests
gunicorn
